Move commons and asserts
