+++ src/lib/EmailInput.js
