+++ src/lib/InputStrings.js
