+++ src/lib/LongError.js
