+++ src/lib/NumberInput.js
