+++ src/lib/SelectInput.js
@@ -10,11 +10,11 @@ import {
   Icon
 } from '@material-ui/core';
 import { KeyboardArrowDown, KeyboardArrowUp } from '@material-ui/icons';
-import { status, isEmpty, isTextLong } from '../../commons/utils';
-import LongPlaceHolder from '../LongPlaceHolder';
-import LongError from '../LongError';
-import { list } from '../InputStrings';
-import '../../styles/BaseInput.css';
+import { status, isEmpty, isTextLong } from './commons/utils';
+import LongPlaceHolder from './LongPlaceHolder';
+import LongError from './LongError';
+import { list } from './InputStrings';
+import '../styles/BaseInput.css';
 
 const useStyles = makeStyles(theme => ({
   root: {
